requirements.txt
streamlit==1.32.0
boto3==1.34.44
pandas==2.2.0
python-dotenv==1.0.1
requests==2.31.0
plotly==5.18.0
fpdf2==2.7.7
beautifulsoup4==4.12.2
selenium==4.15.0
yt-dlp==2023.12.30
lxml==4.9.3
webdriver-manager==4.0.1
